+++ src/main/java/fudan/pbl/mm/domain/Position.java
@@ -1,12 +1,11 @@
 package fudan.pbl.mm.domain;
 
 import java.util.Calendar;
-import java.util.Date;
 
 public class Position {
-    private int x;
-    private int y;
-    private int z;
+    private float x;
+    private float y;
+    private float z;
     private float rotation;
     private int lastRand;
 
@@ -19,39 +18,39 @@ public class Position {
     public Position(boolean random){
         if(random) setRandomPosition();
     }
-    public Position(int x, int y, int z){
+    public Position(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
     }
-    public Position(int x, int y, int z, float rotation){
+    public Position(float x, float y, float z, float rotation){
         this.x = x;
         this.y = y;
         this.z = z;
         this.rotation = rotation;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setZ(int z) {
+    public void setZ(float z) {
         this.z = z;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public float getZ() {
         return z;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
